Rename 'datasets' module to 'data_modules' to fix import conflicts (v0.0.24)
- Rename internal 'datasets' module to 'data_modules' to avoid conflicts
- Update all import references throughout the codebase
- Simplify Colab notebook to use HuggingFace datasets without conflicts
- Update PruningAndFineTuningColab.py to version 0.0.24
- Greatly improve Colab compatibility without complex workarounds
