+++ data_modules/__init__.py
@@ -1,0 +1,3 @@
+# Make data_modules a proper module
+# This module contains data loading utilities for the project
+# Renamed from 'datasets' to avoid conflicts with HuggingFace's datasets package
