--- datasets/__init__.py
@@ -1,1 +1,0 @@
-# Make datasets a proper module
